refactor: change styles for modules
