+++ src/components/Forms/NewPasswordForm.module.scss
@@ -1,6 +1,6 @@
 @import '../../styles/global';
 
-.registration_address {
+.registration_data_name {
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -8,15 +8,16 @@
   color: #333;
   font-size: 2vw;
   padding: 5px;
-  margin: 5px 0;
+  margin: 0px 2vw;
   background: transparent;
-  border: 1px solid $color-brown;
-  border-radius: 8px;
 }
 
-.span_head {
-  color: $color-brown;
-  font-weight: 400;
-  font-size: 1.2rem;
-  margin: 0.3rem 1rem;
+.input {
+  border-top-left-radius: 0.5rem;
+  border-bottom-left-radius: 0.5rem;
+}
+
+.span {
+  border-top-right-radius: 0.5rem;
+  border-bottom-right-radius: 0.5rem;
 }
