+++ src/components/ListAddress.module.scss
@@ -1,0 +1,33 @@
+@import '../styles/global';
+
+.registration_data_name {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+  color: #333;
+  font-size: 2vw;
+  padding: 5px;
+  margin: 0px 2vw;
+  background: transparent;
+}
+
+.list_address {
+  margin: 0.3rem;
+  padding: 0.5rem;
+  background-color: white;
+  border-color: $color-brown;
+  border: 0.05vw solid $color-brown;
+  border-radius: 8px;
+}
+
+.message {
+  font-size: 150%;
+  color: $color-brown;
+  margin: 0.25rem;
+}
+
+.module__window {
+  width: 40vw;
+  text-align: center;
+}
